convert IntComputer to longs

+++ 2019/advent.2019.tests/Day2Tests.cs
@@ -25,7 +25,7 @@ namespace advent._2019.tests
         [Fact]
         public async Task AddAsync()
         {
-            var proc = new Process(new[] { 1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50 });
+            var proc = new Process(new[] { 1L, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50 });
             bool ok = await new IntComputer().ExecuteInstruction(proc);
             Assert.True(ok);
             Assert.Equal(70, proc.Data[3]);
@@ -34,7 +34,7 @@ namespace advent._2019.tests
         [Fact]
         public async Task MultiplyAsync()
         {
-            var proc = new Process(new[] { 1, 9, 10, 70, 2, 3, 11, 0, 99, 30, 40, 50 }, ip: 4);
+            var proc = new Process(new[] { 1L, 9, 10, 70, 2, 3, 11, 0, 99, 30, 40, 50 }, ip: 4);
             bool ok = await new IntComputer().ExecuteInstruction(proc);
             Assert.True(ok);
             Assert.Equal(3500, proc.Data[0]);
@@ -43,17 +43,17 @@ namespace advent._2019.tests
         [Fact]
         public async Task DoneAsync()
         {
-            var proc = new Process(new[] { 3500, 9, 10, 70, 2, 3, 11, 0, 99, 30, 40, 50 }, ip: 8);
+            var proc = new Process(new[] { 3500L, 9, 10, 70, 2, 3, 11, 0, 99, 30, 40, 50 }, ip: 8);
             bool ok = await new IntComputer().ExecuteInstruction(proc);
             Assert.False(ok);
         }
 
         [Theory]
-        [InlineData(new[] { 1, 0, 0, 0, 99 }, new[] { 2, 0, 0, 0, 99 })]
-        [InlineData(new[] { 2, 3, 0, 3, 99 }, new[] { 2, 3, 0, 6, 99 })]
-        [InlineData(new[] { 2, 4, 4, 5, 99, 0 }, new[] { 2, 4, 4, 5, 99, 9801 })]
-        [InlineData(new[] { 1, 1, 1, 4, 99, 5, 6, 0, 99 }, new[] { 30, 1, 1, 4, 2, 5, 6, 0, 99 })]
-        public async Task RunToCompletionAsync(int[] start, int[] done)
+        [InlineData(new[] { 1L, 0, 0, 0, 99 }, new[] { 2, 0, 0, 0, 99 })]
+        [InlineData(new[] { 2L, 3, 0, 3, 99 }, new[] { 2, 3, 0, 6, 99 })]
+        [InlineData(new[] { 2L, 4, 4, 5, 99, 0 }, new[] { 2, 4, 4, 5, 99, 9801 })]
+        [InlineData(new[] { 1L, 1, 1, 4, 99, 5, 6, 0, 99 }, new[] { 30, 1, 1, 4, 2, 5, 6, 0, 99 })]
+        public async Task RunToCompletionAsync(long[] start, int[] done)
         {
             var x = await new IntComputer().RunProgram(start);
             Assert.Equal(done.Length, x.Length);
